backend/requirements.txt
fastapi==0.104.1
sqlmodel>=0.0.16
pydantic>=2.5.0
pydantic-settings>=2.1.0
psycopg2-binary>=2.9.9
python-dotenv==1.0.0
uvicorn[standard]==0.24.0
pytest==7.4.3
httpx==0.25.2
pyjwt==2.8.0
bcrypt==4.1.2
email-validator>=2.0.0
alembic>=1.13.0
apscheduler>=3.10.4
